move PatientPicker to under src/components/Login/
